+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 ## unreleased
 * Add `recurringCustomerConsent` and `recurringMaxAmount` to `authenticationInsightOptions` for `PaymentMethodNonce.create`
+* Add `DOCUMENT_UPLOAD_FILE_IS_EMPTY` error code
 
 ## 5.0.0
 * Add `DISPUTE_ACCEPTED`, `DISPUTE_DISPUTED`, and `DISPUTE_EXPIRED` webhook constants